reorg db helper types
