+++ services/app-api/src/postgres/contractAndRates/prismaSubmittedContractHelpers.ts
@@ -1,0 +1,15 @@
+import { ContractTable } from '@prisma/client'
+import {
+    ContractRevisionTableWithFormData,
+    RateOnContractHistory,
+} from './prismaSharedContractRateHelpers'
+
+type ContractRevisionTableWithRates = ContractRevisionTableWithFormData & {
+    rateRevisions: RateOnContractHistory[]
+}
+
+type ContractTableWithRelations = ContractTable & {
+    revisions: ContractRevisionTableWithRates[]
+}
+
+export type { ContractTableWithRelations, ContractRevisionTableWithRates }
